server/requirements.txt
fastapi
uvicorn
websockets
numpy
mediapipe==0.10.32
opencv-python-headless
tensorflow==2.16.2
supabase
python-dotenv
